add HMAC-SHA1

+++ sha1.c
@@ -261,7 +261,7 @@ void SHA1Final(
     memset(&finalcount, '\0', sizeof(finalcount));
 }
 
-void cal_sha1_hash(unsigned char* hash, const unsigned char* str, int len)
+void cal_sha1_hash(unsigned char* hash, const unsigned char* str, size_t len)
 {
     SHA1_CTX ctx;
     unsigned int ii;
